Chapter 7 - Create an injectable creatable component.
When the Eclipse 4 application runtime content needs to inject objects,
it consults the context chain. If it cannot find an object of the
required type, it will attempt to create an instance of that class as
long as it is marked with the @Creatable annotation.

Thi is used to create a StringService, used by the Rainbow part, to
capitalise the strings passed in.