Chapter 10 - Building an update site
An update site is controlled by a category.xml file, which is the
same content as the site.xml file used earlier.

Note that if the version and url parameters have been set in the
file, then the build may fail to find the exact version of the feature
specified. Resetting it to 0.0.0 will resolve the problem.