Chapter 10 - Using target platform definitions
This adds a target platform definition (based on Juno) which
restricts the available features to the Eclipse Platform, and the
JGit feature which is used in the feature dependencies section.

Running 'mvn clean package' will use the standard Eclipse repository.

Running 'mvn -Dtarget clean package' will use the platform target.