Extras - Using Maven pomDependencies consider
In order to consume bundles from Maven Central (that may not be in
a P2 repository) the dependency needs to be specified in a standard
Maven dependencies block, and the tycho platform needs to include
a configuration <pomDependencies>consider</pomDependencies>.

This adds a dummy class to the clock.ui plugin to simulate a bundle
which is not available in the standard Eclipse Juno/Kepler repositories,
such that it can be compiled and built against.

Note that although the bundle compiles against the package, the
integration test will fail as the target environment does not get setup
with the pom dependency. This must be converted to a P2 repository before
it can be used in the test harness.