Chapter 4 - Creating commands, handlers and menus
A command is a logical action while a handler is a specific response
to that logical action. To integrate with the Eclipse 4 IDE, the
commands and handlers are provided in a fragment file, which is merged
into the workbench upon startup.

Commands and handlers are injected into the Eclipse 4 IDE's commands
and handlers list (known as the feature, in the fragment file).

To demonstrate this working, a simple menu was added underneath the
Help menu in order to allow the command -- and thus, the handler --
to be called.