Chapter 4 - Binding keystrokes to commands
A binding table associates a set of keys with a set of commands, and
is active in a binding context. To use the existing application's
"everywhere" context, it needs to be imported as the dialogsAndWindows
context.

Once the binding table is created, a sequence of keys can be defined
(such as M1+9) which will correspond to Cmd on OSX and Ctrl on other
platforms. When this keystroke is pressed, the command will be executed.