Chapter 4 - Adding pop-up menus to an e4 view
In order to add pop-up menus to an e4 view, it needs to be converted
from an e4view in the plugin.xml to a part descriptor. It is also
necessary to register the menu handler programmatically; once this
is done, it is possible to add pop-up menus through the fragment.