Chapter 5 - Responding to preference changes
Because of the way that E4 injection works, once a value has been
set it doesn't change. This causes problems if the user later comes
along and wants to change their initial value.

By creating a setter method with an injectable parameter, whenever
the E4 preference value changes it will call this method, which can
be used to update the part appropriately.