Chapter 7 - Use an interface for the injectable service
Instead of using a concrete class for the service, an interface can be
used instead. This helps in decoupling the service implementation from
the service usage.

This doesn't work automatically in the injectable case. To fix this, the
injection factory needs to be configured with a binding between the
interface type and the concrete implementation.