Chapter 9 - Adding the CSS spy
Although the product shouldn't ship with the CSS spy enabled, it can
be added to the product and then launched, so that the keypress
Alt + Shift + F5 brings up the spy as well.

When modifying the product, check that all required dependencies are
added, and delete the launch configuration and re-create it so that
the newly added dependencies are present when it is launched.