Chapter 11 - Writing a plug-in test
The plug-in tests have references to the Eclipse run-time platform, and
can be used to test things which rely on an OSGi services runtime being
present.

If this project is run with Run->Run As...JUnit test, the test fails.

If this project is run with Run->Run As...JUnit Plug-in Test, the test
succeeds.