Chapter 11 - Waiting for input
When changes occur asynchronously in the UI, such as when a background
compilation takes place, it isn't possible to always use a change in the
UI to trigger the next sequence of tests. However, there is a way of
waiting until a particular condition appears, which can be used to wait
for the results of any particular operation.