Chapter 12 - Automated building of products
The product and application were previously located in the same
plug-in, but Tycho needs to do different things with an application
and with a product.

As a result, a new project com.packtpub.e4.product was created to
hold the product, and the application is used to build the application
code. The feature was modified to add the e4 application content as well.

Products are materialized with eclipse-repository as a packaging type,
and the application is built as a standard eclipse-plugin.