Chapter 12 - Using a target platform for tycho builds
The target platform provides a restricted view of a remote update site,
which is useful for ensuring that dependencies cannot accidentally
used.

Using the target platform created in Chapter 10, and changing the way
the project is built, it is possible to use that target platform to build
the Tycho build.