Appendix A - Using the service
A factory method can be provided which performs the necessary OSGi
lookup operations from code. This simplifies the case of calling
classes, which may not know that OSGi is actually being used.

By passing in a consumer that needs to use the service, it is possible
to acquire the service for a short period of time, then return it to
the context.