Appendix A - Enable configadmin
By creating a configure method, when this is run with configadmin it
is possible to limit the number of timezones returned by this method.
In conjunction with Apache Felix fileinstall, it is possible to create
a configuration and have that dynamically update the configuration
of the service on the fly.

  java -Dfelix.fileinstall.dir=/tmp/config ...
  echo "max=2" > /tmp/config/TimeZonesProvider.cfg

Install org.apache.felix.configadmin and org.apache.felix.fileinstall from
http://search.maven.org